Extract the entity operations from the gateway
